Remote "summary" task endpoint, split into three
